Drop verbose flag from runner scale set init-dind-externals copy
This follows up https://github.com/github/docs/pull/34015 and resolves the issue brought up in https://github.com/actions/actions-runner-controller/issues/3637

As mentioned in the docs PR as well: Having the verbose flag in there is nice for debugging but it logs a ton of mostly information that isn't interesting most of the time. Plus is slows the cp call down having to output it all. It shouldn't be the default recommended way to run it as such.